requirements.txt
python=3.6.13
praat-parselmouth=0.4.3
fastapi=0.83.0
